data_source_mock/pyproject.toml
[tool.poetry]
name = "data_source_mock"
version = "0.1.0"
description = ""
authors = ["kbpark-onepredict"]

[tool.poetry.dependencies]
python = "3.9"
locust = "^2.5.1"
python-dotenv = "^0.19.2"
rich = "^10.16.1"
numpy = "^1.21.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
ipykernel = "^6.6.0"
wemake-python-styleguide = "^0.16.0"
black = "^21.12b0"
mypy = "^0.930"
MonkeyType = "^21.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
